stock_analyzer/services/ingestion/requirements.txt
requests==2.31.0
pandas==2.2.0
python-dotenv==1.0.1
tenacity==8.2.3
apscheduler==3.10.4
psycopg2-binary==2.9.9